# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Leaflet URL Shortener API
  description: Privacy-first URL shortener backend API
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Configured API server
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
    BearerAuth:
      type: http
      scheme: bearer
      description: OAuth 2.0 access token
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [user, privileged, admin]
        created_at:
          type: string
          format: date-time
    ShortUrl:
      type: object
      properties:
        id:
          type: integer
        shortCode:
          type: string
        originalUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        isCustom:
          type: boolean
        createdBy:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        error:
          type: string
...
paths:
...
  /auth/me:
    get:
      summary: Get current authenticated user
      tags: [Auth]
      security:
        - sessionCookie: []
        - BearerAuth: []
      responses:
        '200':
          description: Current user (with optional scopes for OAuth tokens) or null if not authenticated
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: '#/components/schemas/User'
                      - type: object
                        properties:
                          scopes:
                            type: array
                            items:
                              type: string
                            description: Granted OAuth scopes (only present for OAuth token requests)
                  - type: 'null'
        '401':
          description: Bearer token is invalid or expired
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '403':
          description: OAuth token lacks the required user:read scope
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  hint:
                    type: string
  /auth/logout:
    post:
      summary: Log out current session
      tags: [Auth]
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
  /api/shorten:
    post:
      summary: Create a short URL
      tags: [URLs]
      security:
        - sessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, ttl]
              properties:
                url:
                  type: string
                  format: uri
                  description: The URL to shorten
                ttl:
                  type: string
                  enum: [5m, 1h, 24h, never]
                  description: Time-to-live. "never" is admin only.
                alias:
                  type: string
                  minLength: 3
                  maxLength: 50
                  description: Custom short code alias. Privileged/admin only.
      responses:
        '201':
          description: Short URL created
          content:
            application/json:
              schema:
                type: object
                properties:
                  shortCode:
                    type: string
                  shortUrl:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    nullable: true
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Alias already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (anonymous users)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/urls:
    get:
      summary: List all URLs (admin only)
      tags: [URLs]
      security:
        - sessionCookie: []
        - BearerAuth: []
      responses:
        '200':
          description: List of all short URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShortUrl'
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
  /api/urls/{id}:
    delete:
      summary: Delete a URL (admin only)
      tags: [URLs]
      security:
        - sessionCookie: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: URL deleted
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
        '404':
          description: URL not found
  /admin/users:
    get:
      summary: List all users (admin only)
      tags: [Admin]
      security:
        - sessionCookie: []
        - BearerAuth: []
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
  /admin/users/{id}/role:
    patch:
      summary: Update a user's role (admin only)
      tags: [Admin]
      security:
        - sessionCookie: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [user, privileged]
      responses:
        '200':
          description: Updated user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
        '404':
          description: User not found
  /admin/urls/{id}:
    delete:
      summary: Delete any URL (admin only)
      tags: [Admin]
      security:
        - sessionCookie: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: URL deleted
        '401':
          description: Unauthorized
        '403':
          description: Admin access required
        '404':
          description: URL not found